Migrate Gatsby parts to TypeScript
