+++ package.json
@@ -14,6 +14,7 @@
     "@fortawesome/react-fontawesome": "^0.1.11",
     "@rstacruz/gatsby-remark-component": "^2.0.0",
     "@types/is-url": "^1.2.28",
+    "@types/postcss-custom-properties": "^9.1.0",
     "@types/prismjs": "^1.16.1",
     "@types/reach__router": "^1.3.5",
     "@types/react": "^16.9.44",
@@ -22,6 +23,8 @@
     "@types/react-highlight-words": "^0.16.1",
     "@types/react-instantsearch-core": "^6.3.0",
     "@types/react-instantsearch-dom": "^6.3.0",
+    "@types/sass": "^1.16.0",
+    "@types/twemoji": "^12.1.1",
     "@typescript-eslint/eslint-plugin": "^3.8.0",
     "@typescript-eslint/parser": "^3.8.0",
     "algoliasearch": "^4.4.0",
@@ -87,6 +90,7 @@
     "stylelint-config-recess-order": "^2.1.0",
     "stylelint-config-standard": "^20.0.0",
     "stylelint-scss": "^3.18.0",
+    "ts-node": "^8.10.2",
     "tsconfig-paths-webpack-plugin": "^3.3.0",
     "twemoji": "^13.0.1",
     "typescript": "^3.9.7",
@@ -104,7 +108,7 @@
     "i18n": "extract-messages -o translations -l=ja -d=ja -f yaml 'src/**/*.ts'",
     "lint": "yarn lint:js && yarn lint:ts && yarn lint:css",
     "lint:js": "eslint '{src,plugins}/**/*.js'",
-    "lint:ts": "eslint 'src/**/*.{ts,tsx}'",
+    "lint:ts": "eslint '{,src}/**/*.{ts,tsx}'",
     "lint:css": "stylelint 'src/**/*.scss'",
     "start": "gatsby develop"
   },
