--- plugins/historia-compat-plugin/gatsby-browser.js
@@ -1,7 +1,0 @@
-'use strict'
-
-/* eslint-disable global-require */
-exports.onClientEntry = () => {
-  require('core-js/modules/web.dom-collections.for-each')
-  require('whatwg-fetch')
-}
