+++ plugins/historia-compat-plugin/gatsby-browser.ts
@@ -1,0 +1,8 @@
+import { GatsbyBrowser } from 'gatsby'
+
+/* eslint-disable @typescript-eslint/no-require-imports */
+/* eslint-disable global-require */
+export const onClientEntry: GatsbyBrowser['onClientEntry'] = () => {
+  require('core-js/modules/web.dom-collections.for-each')
+  require('whatwg-fetch')
+}
