--- plugins/historia-compat-plugin/index.js
@@ -1,2 +1,0 @@
-'use strict'
-// Noop
