+++ plugins/historia-feed-plugin/index.ts
@@ -1,0 +1,1 @@
+// Noop
