+++ plugins/historia-soarer-update-plugin/index.ts
@@ -1,0 +1,1 @@
+// Noop
