--- plugins/historia-taxonomy-plugin/index.js
@@ -1,3 +1,0 @@
-'use strict'
-
-exports.createQuery = require('./createQuery')
