+++ plugins/historia-taxonomy-plugin/index.ts
@@ -1,0 +1,4 @@
+import createQuery from './createQuery'
+export {
+  createQuery,
+}
